requirements.txt
aiogram==2.16
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.1
asynctest==0.13.0
attrs==21.2.0
Babel==2.9.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.9
dateparser==1.1.0
frozenlist==1.2.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.2
multidict==5.2.0
mysql-connector-python==8.0.27
numpy==1.21.4
pandas==1.3.4
protobuf==3.19.1
PyMySQL==1.0.2
pyTelegramBotAPI==4.2.1
python-binance==1.0.15
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2021.11.10
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.27
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
ujson==4.3.0
urllib3==1.26.7
websockets==9.1
yarl==1.7.2
zipp==3.6.0
